The tests all passed
I took a different approach here. I've commented out some of the
settings in global_settings.py, and then made the tests pass. I've
already found one instance the tests all passing while a setting is
still being used in the code (databases). So this method is imperfect
but it should give us an idea about what still needs to be worked on.